fix: race condition inside DescendantIndexer

+++ packages/backend/src/peripherals/database/migrations/107_values.ts
@@ -18,9 +18,9 @@ export async function up(knex: Knex) {
     table.dateTime('timestamp', { useTz: false })
     table.string('project_id')
     table.string('data_source')
-    table.integer('external')
-    table.integer('canonical')
-    table.integer('native')
+    table.bigInteger('external')
+    table.bigInteger('canonical')
+    table.bigInteger('native')
     table.primary(['timestamp', 'project_id', 'data_source'])
   })
 }
